Add SMTP config tests
